Remove poke(ushort, MASK) overload — causes ambiguity
The overload causes AmbiguousMatchException in the transpiler's
ReflectionCache AND compile-time ambiguity for integer literals
(0x1E matches both byte and MASK : byte).

Use ppu_mask() for type-safe MASK writes instead of poke().
Where poke() is needed with MASK values, use an explicit (byte) cast.

+++ src/dotnes.tests/RoslynTests.cs
@@ -4206,7 +4206,7 @@ public class RoslynTests
             poke(PPU_ADDR, 0x20);
             poke(PPU_ADDR, 0x00);
 
-            poke(PPU_MASK, MASK.BG | MASK.SPR | MASK.EDGE_BG | MASK.EDGE_SPR);
+            poke(PPU_MASK, (byte)(MASK.BG | MASK.SPR | MASK.EDGE_BG | MASK.EDGE_SPR));
 
             while (true) ;
             """);
